requirements.txt
cassandra-driver
Flask 
flask-cors
astrapy
python-dotenv
datasets 
langchain 
openai 
autogen
Flask
Gunicorn
pinecone-client
supabase